+++ piet-common/src/web_back.rs
@@ -161,7 +161,7 @@ impl<'a> BitmapTarget<'a> {
         let width = self.canvas.width();
         let image = self.raw_pixels(ImageFormat::RgbaPremul)?;
         let file = BufWriter::new(File::create(path).map_err(Into::<Box<_>>::into)?);
-        let mut encoder = Encoder::new(file, width as u32, height as u32);
+        let mut encoder = Encoder::new(file, width, height);
         encoder.set_color(ColorType::Rgba);
         encoder
             .write_header()
